Move resource content preview types to /models
