+++ src/client.spec.ts
@@ -22,7 +22,6 @@ import {
 	MacrosFilter,
 	MyNotificationsMessageReceived,
 	MyNotificationsMessageSent,
-	ResourceContentPreview,
 	ScheduledQueriesFilter,
 	ScheduledScriptsFilter,
 	SearchDownloadFormat,
@@ -70,6 +69,7 @@ import {
 	RemoteKit,
 	RenderModule,
 	Resource,
+	ResourceContentPreview,
 	SavedQuery,
 	ScheduledQuery,
 	ScheduledScript,
