+++ src/models/resource/index.ts
@@ -11,9 +11,12 @@ export * from './is-blank-raw-resource';
 export * from './is-resource';
 export * from './raw-creatable-resource';
 export * from './raw-resource';
+export * from './raw-resource-content-preview';
 export * from './raw-updatable-resource-metadata';
 export * from './resource';
+export * from './resource-content-preview';
 export * from './to-raw-creatable-resource';
 export * from './to-raw-updatable-resource-metadata';
 export * from './to-resource';
+export * from './to-resource-content-preview';
 export * from './updatable-resource';
